api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
python-dotenv==1.0.1
